prevented curse words from being added via shell (Alina's request)
